Split common out in logical pieces
